moving the internal read/write functions to common/ompio and update all fs/fcoll/sharedfp components to use these functions.
