+++ ompi/mca/fbtl/pvfs2/fbtl_pvfs2.h
@@ -23,7 +23,7 @@
 #include "ompi_config.h"
 #include "ompi/mca/mca.h"
 #include "ompi/mca/fbtl/fbtl.h"
-#include "ompi/mca/io/ompio/io_ompio.h"
+#include "ompi/mca/common/ompio/common_ompio.h"
 #include "ompi/mca/fs/pvfs2/fs_pvfs2.h"
 #include "pvfs2.h"
 #include "pvfs2-compat.h"
