+++ ompi/mca/sharedfp/addproc/sharedfp_addproc_iread.c
@@ -9,7 +9,7 @@
  *                         University of Stuttgart.  All rights reserved.
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
- * Copyright (c) 2013-2015 University of Houston. All rights reserved.
+ * Copyright (c) 2013-2016 University of Houston. All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -66,7 +66,7 @@ int mca_sharedfp_addproc_iread(mca_io_ompio_file_t *fh,
                         "mca_sharedfp_addproc_iread: Offset received is %lld\n",offset);
 	}
         /* Read from the file */
-        ret = ompio_io_ompio_file_iread_at ( sh->sharedfh, offset, buf, count, datatype, request);
+        ret = mca_common_ompio_file_iread_at ( sh->sharedfh, offset, buf, count, datatype, request);
     }
 
     return ret;
@@ -180,7 +180,7 @@ int mca_sharedfp_addproc_read_ordered_begin(mca_io_ompio_file_t *fh,
     }
 
     /* read from the file */
-    ret = ompio_io_ompio_file_iread_at_all(sh->sharedfh,offset,buf,count,datatype,&fh->f_split_coll_req);
+    ret = mca_common_ompio_file_iread_at_all(sh->sharedfh,offset,buf,count,datatype,&fh->f_split_coll_req);
     fh->f_split_coll_in_use = true;
 
 exit:
