+++ ompi/mca/sharedfp/addproc/sharedfp_addproc_read.c
@@ -9,7 +9,7 @@
  *                         University of Stuttgart.  All rights reserved.
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
- * Copyright (c) 2013-2015 University of Houston. All rights reserved.
+ * Copyright (c) 2013-2016 University of Houston. All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -62,7 +62,7 @@ int mca_sharedfp_addproc_read ( mca_io_ompio_file_t *fh,
                         "mca_sharedfp_addproc_read: Offset received is %lld\n",offset);
 	}
         /* Read from the file */
-        ret = ompio_io_ompio_file_read_at(sh->sharedfh,offset,buf,count,datatype,status);
+        ret = mca_common_ompio_file_read_at(sh->sharedfh,offset,buf,count,datatype,status);
     }
 
     return ret;
@@ -172,7 +172,7 @@ int mca_sharedfp_addproc_read_ordered (mca_io_ompio_file_t *fh,
     }
 
     /* read from the file */
-    ret = ompio_io_ompio_file_read_at_all(sh->sharedfh,offset,buf,count,datatype,status);
+    ret = mca_common_ompio_file_read_at_all(sh->sharedfh,offset,buf,count,datatype,status);
 
 exit:
     if ( NULL != buff ) {
