+++ ompi/mca/sharedfp/addproc/sharedfp_addproc_write.c
@@ -9,7 +9,7 @@
  *                         University of Stuttgart.  All rights reserved.
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
- * Copyright (c) 2013-2015 University of Houston. All rights reserved.
+ * Copyright (c) 2013-2016 University of Houston. All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -65,7 +65,7 @@ int mca_sharedfp_addproc_write (mca_io_ompio_file_t *fh,
                         "sharedfp_addproc_write: Offset received is %lld\n",offset);
 	}
         /* Write to the file */
-        ret = ompio_io_ompio_file_write_at(sh->sharedfh,offset,buf,count,datatype,status);
+        ret = mca_common_ompio_file_write_at(sh->sharedfh,offset,buf,count,datatype,status);
     }
 
     return ret;
@@ -173,7 +173,7 @@ int mca_sharedfp_addproc_write_ordered (mca_io_ompio_file_t *fh,
     }
 
     /* write to the file */
-    ret = ompio_io_ompio_file_write_at_all(sh->sharedfh,offset,buf,count,datatype,status);
+    ret = mca_common_ompio_file_write_at_all(sh->sharedfh,offset,buf,count,datatype,status);
 
 exit:
     if ( NULL != buff ) {
