+++ ompi/mca/sharedfp/sm/sharedfp_sm_iread.c
@@ -9,7 +9,7 @@
  *                         University of Stuttgart.  All rights reserved.
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
- * Copyright (c) 2013-2015 University of Houston. All rights reserved.
+ * Copyright (c) 2013-2016 University of Houston. All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -78,7 +78,7 @@ int mca_sharedfp_sm_iread(mca_io_ompio_file_t *fh,
 			"sharedfp_sm_iread: Offset received is %lld\n",offset);
         }
         /* Read the file */
-        ret = ompio_io_ompio_file_iread_at(sh->sharedfh,offset,buf,count,datatype,request);
+        ret = mca_common_ompio_file_iread_at(sh->sharedfh,offset,buf,count,datatype,request);
     }
 
     return ret;
@@ -202,7 +202,7 @@ int mca_sharedfp_sm_read_ordered_begin(mca_io_ompio_file_t *fh,
     }
 
     /* read to the file */
-    ret = ompio_io_ompio_file_iread_at_all(sh->sharedfh,offset,buf,count,datatype,
+    ret = mca_common_ompio_file_iread_at_all(sh->sharedfh,offset,buf,count,datatype,
 					   &fh->f_split_coll_req);
     fh->f_split_coll_in_use = true;
