+++ ompi/mca/sharedfp/sm/sharedfp_sm_iwrite.c
@@ -9,7 +9,7 @@
  *                         University of Stuttgart.  All rights reserved.
  * Copyright (c) 2004-2005 The Regents of the University of California.
  *                         All rights reserved.
- * Copyright (c) 2013      University of Houston. All rights reserved.
+ * Copyright (c) 2013-2016 University of Houston. All rights reserved.
  * Copyright (c) 2015      Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
  * $COPYRIGHT$
@@ -80,7 +80,7 @@ int mca_sharedfp_sm_iwrite(mca_io_ompio_file_t *fh,
 			"sharedfp_sm_iwrite: Offset received is %lld\n",offset);
         }
         /* Write to the file */
-        ret = ompio_io_ompio_file_iwrite_at(sh->sharedfh,offset,buf,count,datatype,request);
+        ret = mca_common_ompio_file_iwrite_at(sh->sharedfh,offset,buf,count,datatype,request);
     }
 
     return ret;
@@ -205,7 +205,7 @@ int mca_sharedfp_sm_write_ordered_begin(mca_io_ompio_file_t *fh,
     }
 
     /* read to the file */
-    ret = ompio_io_ompio_file_iwrite_at_all(sh->sharedfh,offset,buf,count,datatype,
+    ret = mca_common_ompio_file_iwrite_at_all(sh->sharedfh,offset,buf,count,datatype,
 					   &fh->f_split_coll_req);
     fh->f_split_coll_in_use = true;
